src/server/mcp_hub/outlook/requirements.txt
fastmcp
httpx
python-dotenv
cryptography
motor
requests
